feat: adjust book title font size based on title and author length 📚
Adjusted the function 'bookFontSize' to take into account the length of both the book title and author. This change will help in better visual representation of the book information. Also, added a new function 'addExtraSpace' to calculate the extra space needed based on the number of words in the author's name.