style: improve typography and adjust CSS styles 🖌️
Updated the font families used in the application for headers and subtitles, changing the previous fonts to 'Lora' and 'Cabin Condensed' respectively.

Adjusted the CSS styles for 'BookCard', 'RankingInfo', and 'Main' components to improve the visual appearance and readability. Modified the grid column width calculation for the 'BookCard' component.

Updated the font sizes in the 'bookInfos.ts' utility file, making the font slightly larger for the same number of letters, which improves readability.

Also, added font preconnects in the 'index.html' file for performance optimization.